fastapi-jwt/requirements.txt
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==3.2.2
beanie==1.10.9
beautifulsoup4==4.11.1
BPMN-RPA==4.8.3
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dependency==0.0.3
distlib==0.3.5
Django==4.0.6
dnspython==2.2.1
docutils==0.19
ecdsa==0.18.0
email-validator==1.2.1
exchangelib==4.7.4
fastapi==0.78.0
fastapi-authorization==0.1.0
fastapi-jwt-auth==0.5.0
fastapi-users==10.1.5
filelock==3.7.1
flasgger==0.9.5
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
functions==0.7.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
jsonschema==4.10.0
keyboard==0.13.5
lockfile==0.12.2
lxml==4.9.1
makefun==1.14.0
MarkupSafe==2.1.1
mistune==2.0.4
mongoengine==0.24.2
motor==2.5.1
MouseInfo==0.1.3
multidict==6.0.2
ntlm-auth==1.5.0
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
pid==3.0.4
platformdirs==2.5.2
pluggy==1.0.0
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycamunda==0.6.1
pycparser==2.21
pydantic==1.9.1
PyGetWindow==0.0.9
Pygments==2.12.0
PyJWT==2.4.0
pymongo==3.12.3
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypi==2.1
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
pytest==7.1.2
python-daemon==2.3.1
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
requests==2.28.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
rsa==4.9
schemas==0.7.1
service==0.6.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
sqlparse==0.4.2
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
uvicorn==0.17.6
validators==0.20.0
virtualenv==20.16.2
watchfiles==0.16.1
websockets==10.3
Werkzeug==2.2.2
xmltodict==0.13.0
yarl==1.8.1
